Perform restraint COM calculation outside of read_posres()
This change is needed to resolve #5126.
The calculation of the COM of restrained atoms in `src/gromacs/gmxpreprocess/grompp.cpp` has been split into 3 auxillary functions (`calc_posres_com()`, `sum_com_for_restraints()`, `sub_com_from_restraints()`) and now it is not performed in `read_posres()` anymore.
The plan is to generalize these auxillary functions to compute a list of COMs instead.

Closes #5126