Add array reference for COM restraint indices to listed forces
An array containing the COM group index of each atom is needed in order to scale the reference coordinates w.r.t multiple centers of mass.
This merge request simply adds a `restraintComIndex` variable to `ListedForces`. At the present moment this is simply initialized with the group index for center of mass motion removal. It will be needed later in `posres()`.
Required to solve #5126