SYCL: Fix memory corruption with VkFFT 1.3.5
We manually set the pointer in VkFFT's internal datastructure.
Since version 1.3.5 (9a96811), VkFFT now has ownership of these pointers
and tries to `free` them on shutdown, which does not work well with a
pointer to a long gone stack variable. Now we set the pointer to
`nullptr` to avoid it being used after it's gone out of scope.

Not needed for HIP, since it does not do this funky business, having
guarantees of always using the same underlying hipStream.

Fixes #5184