Update cmake and pyenv in plain llvm containers
Had to update cmake and pyenv versions, so also updated llvm and
ubuntu versions.

Eliminated the unnecessary dedicated container for the simple build.

Minor modification to GoogleTest to work around a limitation of
clang-19 in the context of C++20.

Bumped minimum required pybind11 because 2.6 doesn't compile with
clang-19.

Suppressed a compilation warning in googletest via sample_restraint
just like we do for the main build of googletest.

Refs #5014