Change QMMMTopologyPreprocessor member functions to protected
This is the second of several MRs related to the implementation of an interface for neural network potentials described in #5039. This minor change declares the member functions of the `QMMMTopologyPreprocessor` that modify the system topology as protected (as opposed to private). This way, classes that inherit from it can make use of its functionality more flexibly, as needed e.g. by `NNPotTopologyPreprocessor` that will be added in an upcoming MR.