Enabled keyword init-histogram-counts to enable continuation of expanded ensemble runs
This keyword enables the histogram count of expanded ensembles to be initialized at the start of a simulation via the .mdp.  This has ended up being necessary for long expanded ensemble runs locally and for researchers using Folding@Home. We both have been using locally modified branches, but it would be good to finally get it into the main branch of the code.

This should have minimal interaction with any other code paths or functionality.

There may need to be tests written for this - if someone can point me to an appropriate similar test (i.e. add a new keyword to the .mdp and verify the proper data locations are written to), I will do that.