Made OptionsAssigner error message more helpful
When using a .tpr made from a branch with extra mdp options with a
branch that is unaware of those options, the KVT cannot be read and an
error must be given. Adding the name of the options section that
cannot be read to the error message will make it more clear to the
user that they have mismatched .tpr and code.

Refs #5194