Double check that GPU-aware MPI will work
On SYCL devices with Intel MPI, GROMACS already prefers the L0 backend
where both are available and detects GPU-aware support only when L0 is
used. But if L0 backend is not available *and* the user forced
GPU-aware support, crashes result when OpenCL pointers are used in L0
contexts. This check helps the user do a better thing.

In other cases, do nothing.

Refs #4638