Revert "Move PME GPU force receiver reallocation"
This reverts commit c14bda28b97a43dcd19bb55b4f38d02d1dc31488.

With thread-MPI direct communication the remote PME rank x pointer needs
to be received on the PP rank, but this is only available after the
PME-side reallocation happens. Since the repartitioning-related size
changes are mixed with actual data communication on PME ranks as well in
gmx_pme_recv_coeffs_coords() this prevent moving the force receiver
reallocation.

Fixes #5221

Closes #5221