Don't ignore frame read error in gmx tcaf
Return value of `read_first_frame` was not checked in `gmx tcaf`, such that error there resulted in confusing failure of the computation downstream of this call. This was notably the case for the reproducer files in  #5238, which did not contain the necessary velocities.

Now, the return status is checked and an error is emitted.

(Note, there are dozens of this same pattern of ignoring `read_first_frame` in the old `gmxana` tools, a possible follow-up but IIRC policy is not to touch those tools too much so ¯\\_(ツ)_/¯)


Fixes #5238