Small updates of Colvars library; add libTorch-Colvars interface
This MR includes small fixes and improvements to the copy of the Colvars library in `src/external`, as well as one feature (the `torchANN` collective variable type).

The Torch-Colvars interface was previously not included in !4611 because of failures in the Colvars CI runners. We have since confirmed with the main author of the feature that the culprit was the precision of the GROMACS build and he confirmed that the numerical results are consistent across libTorch versions (see https://github.com/Colvars/colvars/pull/570#issuecomment-2476421289).

Matches [this commit in the Colvars repo](https://github.com/Colvars/colvars/commit/3023d8ee56e24a3a620e17452753802326877140).

CC @HubLot @jhenin