Avoid creating std::string on every call to checkDeviceError
`checkDeviceError` and `ensureNoPendingDeviceError` in CUDA and HIP were
taking the error message as `std::string`. This means that every time
they were called with a string literal, a `std::string` object could be
constructed and destructed, potentially with memory allocations, even if
we're on the happy path and no error is to be thrown.

This overhead can easily be avoided by switching to `std::string_view`.

Hat tip to @vedranmiletic for bringing this up in #5254.