Restored xstart_ member of AbstractAnalysisArrayData
Commit e1a21fd1685 introduced the notion of a non-uniform X axis, and
removed xstart_ in favour of sometimes using the first x value to
contain it. This was probably less clear and does make it harder to
refactor the code to use ranges.

This commit restores xstart_ and uses it as previously.

Added more comments about data members, preconditions, and
postconditions.

Removed unused xstart() method.