Forcefully set CMake variable GMX_USE_HDF5 to OFF
The code that uses HDF5 is already in, but not used. Having the CMake
variable GMX_USE_HDF5 set to ON (e.g. inherited from a previous cache)
will result in linker errors.

This change forces the variable to always be OFF in (for release 2025)
and overrides the value stored in CMake cache, if any.