Add gmx_inline and use it in the nbnxm kernels
Add gmx_inline to be used in place of plain inline to expand to compiler-specific
function attributes that force inlining.

Also added missing attributes to the `gen*Arr()` functions in gmx_simd.h which
was part of the reason for some compilers (in particular gcc<13 not inlining).

Performance verified on x86 and Aarch64 / NEON. In-depth testing done on Zen2/Zen3
using gcc 11,13,14 and clang 17,18,19 with a range of kernel flavors (ref #4752).
Performance improves in most cases by a few %, with gcc 11 (most affected) by up to 7%. 

When comparing to GROMACS 2022, the performance regression reported in #4752 is resolved
in gcc, but clang performance is still up to 20% worse.

Refs #4752