SYCL: Add support SYCL_EXT_ONEAPI_ENQUEUE_FUNCTIONS
The DCP++ extension is experimental and, in practice, unstable, so
the feature is opt-in, requiring `-DGMX_SYCL_ENABLE_EXPERIMENTAL_SUBMIT_API=ON`
at configure time.

No major performance gains on RTX3060, but for 1k steps of a PME Grappa test,
reduces the number of API calls quite a bit, although there's still room
for improvement:

| API         | before | after | native CUDA |
| ----------- | ------ | ----- |------------ |
| EventRecord | 21770  | 14125 | 3060        |
| EventCreate | 17778  | 7081  | 68          |

Getting this code in to simplify future testing and remove the
need to maintain a custom branch.