Adding missing defgroups for doxygen
In order to update the Doxygen version, we better reduce warnings. Nearly 385 warnings belong to this category. We used group tags with `\ingroup` tag, without defining them with `\defgroup`. This MR includes defining missings `\defgroup` tags. The following groups are defined (without description):

```
- module_applied_forces
- module_compat
- module_fft
- module_fileio
- module_gmxlib_nonbonded
- module_preprocessing
- module_gpu_utils
- module_hardware
- module_math
- module_mdlib
- module_mdtypes
- module_mimic
- module_pbcutil
- module_pulling
- module_topology
```

I tried to put the definitions as close as possible to their files.

Refs #4543