To bb or not to bb, that is the QUESTION!
As the great bard once wrote:

"To be, or not to be, that is the question."

And here we have its C implementation:

```c
#define QUESTION ((bb),| !(bb))  
```

But alas! A tragic flaw! A syntax error most foul! 


With but a small correction, our existential pondering is restored to its rightful form:

```c
#define QUESTION ((bb) || !(bb))
```

Now, QUESTION is eternally true, as if Shakespeare himself had written it into the very fabric of computation!