Use CMake target-style MPI dependencies
The former global flags are a legacy from CMake 2.6 when we did not use
CMake targets. One should generally use a wrapper compiler and the
CMake target, and as such have no need of appending to global flags.

Likewise when using a CMake target, the MPI target should be added
on an as-needs basis.

There's no need to add MPI C flags