Use const reference to MDLogger in applied forces modules
GROMACS always provides a reference to an MDLogger object, which is
configured with streams that may or may not actually write output.
Unfortunately the callback model used for notifications requires
two-phase initialization of options objects, so the logger starts
life uninitialized. The most natural way to express a handle
that has not yet been initialized is a pointer. We should not
continue to use it as a pointer, because that implies that being
uninitialized is an expected state during normal operations. Our
modules and tests should all work the same way, with an initialized logger.

Introduced getters for the actual reference along with assertions that
the pointer is in fact valid at the time of use, because if invalid
then either the getter was called too soon or the notification for the
MDLogger is missing/broken.

Minor related improvements to variable naming and Doxygen.