Create serialization module
Utility module functionality should have no other GROMACS
dependencies, but serialization depends on RVec which is in the math
module and also utility module.

Removed existing workaround for rvec for ISerializer interface in
favour of depending on the definition in math module, so that we can
also do that for RVec later on.