Show GoogleTest skip messages also with library MPI
Long ago, we implemented support for delaying output from GoogleTest
binaries run with an MPI library until the end of each test, so that
the output from multiple ranks was easier to understand. We forgot to
report any message output by skipped tests, because we weren't using
any skips at the time. Now the skip messages are reported.

Minor improvements to comments