Provide big 5 for IMdpOptionProvider to simplify tests
The original design of IMdpOptionProvider didn't envisage the
level of detail we've adopted in applied-forces tests. The first
applied-forces tests for the electric field module built the whole
module, so the utility of treating the IMdpOptionsProvider object
as a value type did not arise.

We now have several groups of tests that are more fine-grained, which
all had a duplicated helper method that ran the intrinsically complex
mdp-handing workflow, as-if by grompp. The helper method was extra
hard to understand because the IMdpOptionsProvider was a member
variable of the test fixture, rather than created and returned by the
helper method. That object was a member partly because it was not able
to be copied or moved.

Extracted the common helper method as a template and re-used it in
several MDModules tests. This provides a common point to document
the workflow clearly.

There's no changes to test coverage or execution workflow.

Moved constant strings for several modules to the implementation file,
which reduces the scope to the necessary level and permits
implementing the big 5 for the respective options classes. This was
probably the original design error that prompted the first
implementation of setFromMdpValues().

Removed some unused member variables, related functions, and include
files. Made function names conform to style. Cleaned up some use of
namespaces. Permitted QMMM parameter class to be movable and
assignable because there is no need to restrict that for types that
are already value types.