Stored MDModules as a map of named modules
This decreases the coupling between class MDModule and the classes
that implement IMDModule. We can refer to modules by the value
of a string, check that duplicate modules are not added.

Updated some high-level docs and TODO items.

Used ModuleInfo classes consistently.

Used constexpr string_view consistently for module-name strings, and
propagated that into the checkpointing interface, which is now less
opinionated about the string identifiers that it needs.