Fixed execution of tests with cross-compiling emulator
When CMAKE_CROSSCOMPILING_EMULATOR is set, the tests using generator
expressions, such as $<TARGET_FILE:helloworld>, will ignore the
setting. Now we explicitely prefix the test commands executed by CTest
with the executable in CMAKE_CROSSCOMPILING_EMULATOR, if any.

This eases the testing of GROMACS for Windows on Linux with Wine in
situations where no special configuration for Windows executables is
present in binfmt_misc, which seems to be the default.

Fixes #5342