docs: Improve install guide for Intel Xe2
The devices only support sub-groups 16 and 32, so cluster size of 8
should be used. No data about `GMX_GPU_NB_NUM_CLUSTER_PER_CELL_X`,
but for simplicity let's not create a separate entry.

Also avoid suggesting specifically cluster size of 4 when we only know
that it's incompatible. Could be made even smarter, but for the release
branch let's at least make sure it's not misleading.

Refs #5330