CI: Work around timeouts on AMD GPUs
Two improvements to the CI

1. Reduce `GPU_MAX_HW_QUEUES` to `2` (default is 4).

RX6400 we have in CI only has 8 HW queues available to ROCm, and when running
multiple tests in parallel, oversubscription can happen.
Less of a problem for MI50 (24 HW queues available to ROCm), but still lets
be consistent.

2. Disable SDMA on gfx1034.
This improved situation in manual testing, but the mechanism is less understood.
Thanks to Stefan Fleischmann (@stiobhan) for finding this.

Refs #5341