Add extra tests for pdb2gmx handling of cyclic peptides
There was a bug when a cyclic peptide was not first in the chain
that is now coverered.

Also, when multiple chains are present and not merged they are written
to per-molecule .itp files, which were not compared with reference
test data. Now they can be.

This required two updates to cmdlinetest infrastructure: Sometimes
GROMACS preprocessing and analysis tools generate output filenames
from raw strings, and sometimes by modifying names of files supplied
on the command line. In the latter case (used here by pdb2gmx), these
are prefixed with the name of GoogleTest temporary
directory. However the label of the data set in the reference data
file should not have that prefix. So setOutputFileWithGeneratedName
now allows the caller to choose whether the label and prefix are the
same or not, according to how the GROMACS tool functions.

Updated to use string_view and filesystem while I was there.

Modified tests for gmx chi to use the expanded approach.

Covered per-molecule .itp files written by the existing cyclic RNA
test input.

Added two new test cases to cover the buggy cases with multiple cyclic
peptides perhaps of different sizes or following a non-cyclic chain.

Refs #5091