Add IMPI to oneapi images and bump to 2025.0 and 2025.1
IMPI support requires use of Intel APT repos from multiple
stages, so extracted a helper function for that logic.

Codeplay's plugin download changed approach for oneAPI 2025.1,
so introduced new logic to work with that.

Used oneAPI 2025.1 for the testing on Intel GPUs so that adoption
of distributed GPU support in MKL and device-initiated communication
will have the best support available. Used oneAPI 2025.0 for testing
the SYCL build on other hardware.

Replaced OpenCL-on-Intel configuration with
SYCL-and-IMPI-on-Intel to prepare for the above adoption, using the
Level Zero SYCL backend (as required for that).

Disable MPI support from python env for simplicity

Refactored global application of floating-point flags so that they
are not used on targets that don't need them, which helps avoid
linker warnings about the use of unused flags.

Ubuntu 24.04 container bases do not build smoothly for now