Separate initial behaviours related to log file
The fatal error handler would like to write output to both the
terminal and to the log file, but it can only have the latter when the
log file is open for writing, which happens at a different point in
mdrun according to whether output files are being appended.

Previously this was bundled alongside the code that writes the binary
information to the header, but that was more of a complication than
duplicating the call to the fatal-error-handler setter.

Now used sg::scope_guard for clearing the possible log-file handler
held by the fatal error handler, rather than relying on
closeLogFile. This is better separation of concerns.