Remove per-object locking from t_fileio
libgromacs maintains a per-process doubly linked list of open t_fileio
handles, so that checkpointing can store the state of those file
pointers at the time the checkpoint was written, so that restarts with
appending can do the right thing. This dates from the C era of
GROMACS, and is done with an intrusive linked list, ie. t_fileio
contains pointers that to point to the nodes before and after it in
the list.

There's a "dummy" head node of the list stored in a file-static
variable. In principle, threads could want to traverse
the list to find particular files concurrently with other threads
opening and closing files (which thereby modify the list). It's hard
to make that correct and performant while permitting fine-grained
access to the list. Since GROMACS doesn't need fine-grained access
of that type, it uses a file-static mutex 'open_file_mutex' that
ensures that no two threads operating on the prev and next fields
concurrently access the list. After mdrun startup, such operations
only happen during checkpointing on the main rank, so this is not
a performance issue.

This MR affects none of the above.

Once a thread has access to a t_fileio object it can want to do things
with it, like read/write data from the filehandles, get values from
the object, or do file operations like seek, tell, truncate. In
principle, multiple threads could want to do such operations
concurrently with eachother (or with list-traversal operations). By
design, GROMACS does not do such things e.g. even with thread-MPI, the
main rank does all file I/O; debug output goes to per-rank files;
trajectory analysis is single-threaded.

None of these operations have an intrinsic need for mutual exclusion
with another thread doing a similar such operation on the same
t_fileio object. Formerly, gmxfio.cpp used a lock for each t_fileio so
that such operations were atomic ie. no thread ever saw the unfinished
work of another thread. That was false safety, in the sense that such
correctness of such code needs higher-level coordination (so that
e.g. no thread is flushing while another is truncating the same file,
or no thread is asking for the file position while another is closing
it or writing data to it). Without such coordination, the result
depends on the undefined behaviour of the race between the two
operations, regardless of whether the operations were themselves
atomic. This MR removes the atomicity, which only created the
appearance of correctness. Correctness in the presence of
multi-threading cannot be assured merely by low-level mutual
exclusion.

This MR thereby relaxes an unnecessary dependency on thread-MPI. It
might make some kinds of file operations a tiny bit faster.

Removed declaration of gmx_fio_fsync, which was unused outside its
translation unit.

Noted where responsibility for flushing TNG output files before
checkpointing actually lies and documented presumed limitation

Refs #5358