Let t_trxstatus store the type of its file
Previously this was tracked by t_fileio, but TNG files were never
supported by that (probably because t_fileio was too complicated to
change safely). Noted one such deficiency. Now t_trxstatus keeps track
of the file type itself, which makes sense given that most things it
does have to switch on the type of the file. This also reduces the
need for use of getters and local variables, and the amount of
special-case code, overall simplifying the code.