H5md: Add reading and writing of lists of RVecs
This commit adds creation of HDF5 data sets for per-frame lists of `gmx::RVec` objects as well as template specializations of `readFrame` and `writeFrame`.

The data type for these data sets are 2D arrays with `[numAtoms, DIM]` dimensions and `real` type. After creating these data sets the array size is fixed, but the data set can be resized along the frame axis to allow for reading and writing of trajectory data such as positions and velocities.

Tests are added to ensure that the data sets are of correct size and type, that reading and writing buffers of `ArrayRef<RVec>` type can be done to any available index in the data set, and that these buffers are of the same size and type as the arrays of the data sets: Barring bugs in these checks we will not read from or write out of our allocated memory.

Refs #5016