Unit tests for printing routines
On POSIX systems, at least, we can have unit tests for routines that
print to a FILE*. Introduced a class for use in testing that wraps the
use of open_memstream (if found on the system). On other systems, the
tests still run, but the output goes to stdout and is not inspected
for correctness.

This change (and its test) is a precursor to
replacing these methods with some that use TextWriter for better
indenting, line wrapping, and testing support.

Also fixed interface and location of implementation of gmxSetenv