Let XdrSerializer handle its own files
This frees XdrSerializer from the unnecessary dependency on t_fileio
(along with its reference counting and file lists).

Once checkpointing always uses the ISerializer interface, we can also
hide the XDR dependency in the source file for XdrSerializer, but
for now we have to expose the XDR symbol and a getter.