Remove deep assertions in checkpointing
When reading, writing, or dumping, the state flags are used to
recognize which fields of the simulation state should be
handled. These are serialized first, and used to dispatch the rest of
the checkpointing code. The mid-level checkpointing code loops over
the set of possible flags, checks whether that state-flag bit is set,
and then calls a low-level function to do the work, passing along that
state-flag bit. That state-flag bit value is used to provide a
description when dumping.

The state-flag bit was also used to assert that the value is
consistent with the set that matches the template argument of the
low-level function. That was just what the caller did in all cases, so
was more bureaucracy than value. The bitfield values also overlap
considerably between different sets of state-flag enums (e.g. compare
StateKineticEntry and StateEnergyEntry), so checking the value only
prevented some mismatches. This code will shortly get simpler still
and the value decrease further.

For do_cpt_state a copy of state->flags was always passed from its
callers, which was an additional redundancy.

For doCptPullCoordHist and doCptPullGroupHist, a complete set of flags
was always constructed and then a loop over all flag bits was used to
dispatch the work. This can now be recognized as overly complex and
simplified.