Reorder nd_range indices to launch SYCL NBNxM kernels along dim 2
Reorder indices for nd_range used for launching the SYCL NBNxM kernels,
so that workgroups are indexed along dimension 2 instead of dimension 0.
This change makes the NBNxM launch configuration consistent with the rest
of the SYCL kernels and improves launch behavior by using configurations
that permit launching more workgroups.

For instance, on HIP/CUDA backends, nd_range dimensions are often
limited to 2^16 workgroups in the 0th and 1st dimensions, while in the 2nd
dimension, 2^32 workgroups can be launched.

This change did not result in performance change on MI210, PVC1550, or
PVC1100.