Start ITT profiling data collection after reset
Intel's tracing technology can be configured with data collection
paused from the beginning. If so, it's convenient to start collecting
data when GROMACS counters are reset. This can help align the
conclusions from the two data sets. If data collection was not paused
at the beginning, the call to resume collection has no effect.