Fix Clang 20 build with _LIBCPP_REMOVE_TRANSITIVE_INCLUDES
Clang 20 with libc++ and -D_LIBCPP_REMOVE_TRANSITIVE_INCLUDES is
stricter than the previous versions so it does not define the
__cpp_lib_filesystem symbol unless the <filesystem> or <version> header are
included (and the <version> header is limited to C++20 or newer). Since
the purpose of checking this symbol is to decide whether or not to
include the <filesystem> header, this becomes a chicken and egg problem.

This patch reuses MSVC logic of checking for the existence of the header
for all compilers that don't define the symbol, including (but not
limited to) Clang 20 with libc++.

Refs #5396