Remove parallel run conditions from checkpoint
These were used to provide log output of a mismatch in parallel run
conditions when using mdrun -reprod. This was useful because if those
conditions are correctly reproduced then in principle mdrun could run
reproducibly (except as limited by other aspects, such as GPUs,
dynamic load balancing, or floating-point accumulation). This check is
potentially useful when debugging. However, in such cases there is
generally information in addition to the checkpoint file to aid in
such reproduction, e.g. a log file, run script, or bug-report message.

The feature got broken when checkpoint reading was moved before DD
setup, so that the intended PME-rank count was not available when
running the check. For example, mdrun -reprod logged about a mismatch
whenever the checkpoint was written by an mdrun that used PME ranks,
whether or not there was going to be an actual mismatch.

The old implementation was not particularly complex, but it created a
dependency on the order in which operations occured. Fixing the bug
would mean storing the data read from the checkpoint until after DD
was set up, and that is not worth the minor convenience of the extra
check in mdrun -reprod. Instead the check is removed.

The parallelization data can still be retrieved with `gmx dump`
if it is needed.

Refs #3309