CI: Set GPU_MAX_HW_QUEUES=2 for regression tests
Setting `GPU_MAX_HW_QUEUES=2` limits each process to
four HW queues (two for normal-priority streams and
two for high-priority streams), which is okay for
two processes per gfx1034, which has 8 hardware
queues available to ROCm.

We set it earlier for unit tests (where we run several
tests in parallel), but not for regtests.

Refs #5341