Make BUILD_SHARED_LIBS=off work with SYCL build
When libgromacs is a static library, the metadata to link to libsycl.so
is not available when linking SYCL-enabled libgromacs into a binary.
By adding SYCL support to the binary target, the linking will work
correctly.

Now gmx is built with SYCL in the case where it was needed. The test
binaries with source files needing GPU compilation were always linked
with SYCL, so didn't have this problem. But other test binaries and
the template need the same fix.

Fixes #5393