Change nbnxm default to analytical Ewald on AMD gfx94x
Testing has showed that in the majority of the use-cases, the analytical Ewald kernels is faster than the tabulated.

On MI300A nbnxm kernel has complex size-dependent behavior due to which, of the more common kernel flavors, for Ewald with LJ force-switch this chance introduces a regression. Follow-up changes will try to tune heuristics to avoid that.

Refs #5329