Fix x-updated-on-device event accounting
When using separate PME ranks with PME on a GPU and sending from a
GPU, empty domains never sent a message and never consumed this device
event. This lead to a fatal error in builds that checked device-event
consumption (e.g. SYCL).

Now we expect that event to be consumed in that way only on non-empty
domains.

Refs #3988 
Refs #5132