Use MessageStringCollector in modular simulator compatibility checking
This refactoring separates the building of a set of descriptive error
messages about why the modular simulator cannot be used from the
decision about what to do about any incompatibility. It also avoids
exiting after just the first reason for incompatibility.

This MR changes no behaviour while preparing for an enhancement to log
why the modular simulator cannot be used, and adding full context to
the error given when the user required the modular simulator but the
inputs are incompatible with that.

Refs #5339