Log reasons why modular simulator cannot be used
Users of the velocity-Verlet integrator might want to know the reasons
why the modular simulator is not being used, and now mdrun logs those
details. When mdrun was required to use the modular simulator and it
is incompatible, it now throws and describes the reasons in detail.

It continues to assert when double-checking consistency when actually
building a modular simulator.

Fixes #5339

Closes #5339