Fix empty domains with separate PME-on-GPU ranks
With pipelining and PME on a separate GPU, the main PME
stream needs to wait on the work in a pipeline stream, but
should only do so when the pipeline stream did work for a
non-empty PP domain. The proposed solution minimizes new
host overhead before launching all kernels.

PmeCoordinateReceiverGpu now keeps track of PP ranks that sent work.
This slows down the normal path further by the cost of at least one
std::vector.clear() and one std::vector.push_back() per step in
order to cater for the unusual cases where one or more PP domains
does not contribute work. But now the dependency accounting is
correct.

Refs #3988
Refs #5132