Reduce DD halo with direct communication and update groups
With direct halo communication (not enabled by default), the atom
group radii were accounted for twice: once in the cut-off and also
in the bonding boxes of grid cells and clusters. This caused the halo
volume to be unnecessarily large with update groups.
Now maxAtomGroupRadius is removed from the target domain grid bounding box
and two different cut-off are used to avoid double counting.

Also corrected the naming of the cut-offs.