Only select cells in range with direct halo comms for CPU
We used to select one cell range to communicate for each grid column:
one between the first and last cell that are in communication range.
This led to an increase in halo size. Now, for CPU grids, we only
communicate cells that are in range.

For direct communication of GPU grids this optimization will be added later.