Make PME grids reallocatable
Due to changes in thread domain decomposition of the rank-local PME grid, the thread-local grid size can increase during PME tuning. This would cause an assertion to fail. Now we don't store a direct pointer to the memory, but rather a pointer to the vector, so we can reallocate the memory when needed. \
    \
Fixes #5406 \
Fixes #5417