Set minimum steps per tau_p to 25 for C-rescale
The C-rescale barostat gives too large fluctuations when tau_p is smaller
than 25 integration steps per period (introduced in the 2023 release).
The minimum number of steps per tau_p is now increased from 5 to 25.
This was only an issue when tau_p was smaller than the recommended 5 ps
for simulations with dt=2 fs.

Fixes #5425