Make AWH equilibration check default and increase the tolerance
The default tolerance value of 0.2 for AWH histogram equilibration check
was too small and slowed down convergence in many cases. Now this value can
be set by the user and the default value is increased to 0.5. This significantly
improves convergence and also doesn't seem to affect the convergence with a single
walkers, so it the equilibration check is now turned on by default.