Direct-GPU PP-PME communication always transfers x and f
It was confusing and bug-prone to skip transfers of x or f between PP
and PME ranks for empty domains, because the responsibility for the
check was split across several levels. In normal simulations, domains
are never empty, so we can be both simple and fast by always doing a
transfer, even of an empty range. In particular, we resolve the
situation where with library-MPI, the direct-GPU communication path
did not post receives for domains with no particles.

Note that an MPI operation completes when the caller can re-use the
resources (ie. when a blocking send/recv/wait call returns). This can
be trivially true for a message sending an empty buffer.

Note that while this commit is believed to fix the above issue, this
commit is also known to produce or expose failures or segfaults in
MdrunMpi4RankPmeTests and a real simulation with an empty domain. The
cause of those problems is not known. Those problems appear to be
fixed in the following commit "Fix PME spread event-consumption
expectations."  On its own, that commit is believed to fix the stated
issue, but exposes other problems, so we have incorporated the two
fixes as successive patches and observed that the CI pipelines pass
when both patches are present. Keeping the logical fixes separate will
make future investigation of that possible unknown bug easier, if
necessary.

Fixes #5446