Check for stable CMake behaviour where possible
In many GROMACS build configurations we can and now do check in CI
that a non-initial run of CMake in the same build tree
does not change the values of the CMake cache variables
and does not write status output. The former helps ensure
that our build configuration system does not change how it
builds GROMACS according to the number of times CMake is called.
The latter helps developers notice problems when making change to CMake code.

Some exceptions are an NNPOT build (because PyTorch CMake is irrepressibly
talkative), HIP builds that use VkFFT (HIP CMake is hard to work with),
and an AdaptiveCPP build (which needs simplification after support for
hipSYCL is removed).

To achieve this, the CMake code for several components is improved
in ways that directly achieve the above goals, including for finding
LaTeX, ImageMagick, CUDA Toolkit, HIP, Python3, PkgConfig, and OpenCL.

The GMXAPI option is now declared earlier so that it can influence the
finding of MPI libraries on the first run of CMake.

Some recently orphaned NVCC config code was removed because it made
a cache variable that changed despite never being used.

Refs #5395 #4716