H5md: Re-add setAttributeVector overloads for string types
These were removed (!5298) in favor of `setVectorStringAttribute` which allows us to build a buffer of the written data. This is not required for all string-vector attributes written when an `H5md` file is being set up. This commit re-adds the overloads for `setVectorAttribute` for `ArrayRef<const string>` and `ArrayRef<const char* const>` which covers common use. Tests are added to ensure compatibility with relevant use cases.

Refs #5016