Move gromacs:gcc-11-hip-622-mpi:test to post-merge
It's pretty flaky and likes to lock-up CI resources.

Hopefully won't be the case of "out of sight - out of mind".
We want to unblock the CI, and the risks of breaking this config
are not too high (we still have the build and the regression tests
in the pre-merge).

Refs #5341