Allow PME-only rank to treat direct-GPU-comm setting as constant
This value is constant in the simulation workload on all ranks, so the
PME-only rank can keep its copy from setup time rather than expect to
get it regularly in MPI control messages. It was only being sent in
some control messages, which made it easy to break the PME-only rank.

Note that whether PME ranks return forces direct to GPUs is handled
separately, and that can vary with each simulation step (principally
if it is a virial step).

The name used in pme_pp struct is now consistent with that used in
SimulationWorkload.

Refs #5132 
Refs #5446