Add trajectory reading framework to H5md class
Adds `readNextFrame` method to the `H5md` class, which reads data into `t_trxframe` objects. Extends the `TrajectoryReadCursor` implementation object with methods to select which trajectory data sets (positions, velocities and/or forces) are available for the next frame, and to read that data from our stored data blocks.

Adds tests to ensure that trajectory data is read in the correct order for different output frequencies of positions, velocities and forces.

Refs #5016