Fmm: Fix incorrect direct-provider selection in FMM MDP options
There is a bug in selection of Mdp options, Direct Provider always takes default value because `FmmMdpOptions` initializes a lookup in the constructor based on `exafmmOptions` and `fmSolvrOptions_`. Although both can take different values for direct provider later on but it doesn't get reflected in the lookup.

I identified it when testing latest changes in the interface with ExaFMM.