CUDA: more conservative estimate of max. threads/SM
Previously, we were using 2048 as the `GMX_CUDA_MAX_THREADS_PER_MP` for
unknown architectures. This has been causing issues (#5366, #5479,
5b720497, fdfa9a11, c2e18235): when the value is too high, the build
fails. Inverting the condition so that all unknown architectures will
have the limit of 1536 threads/SM. This can lead to lower performance
of the Prune and Gather kernels, but is better than GROMACS not building
at all.

This does not address the issue of, e.g., using a `compute_90` PTX to
target a 12.0 device: that would still fail.