Permit optional use of OneAPI extension work_group_static in all SYCL kernels
This change may produce a minor performance improvement itself, but
its main role is to prepare for a future change for SYCL kernel
submission that does not use the command-group handler. Not making the
handler is expected both to improve both performance and reduce performance
variability.

Accordingly added CMake support via a top-level flag
GMX_SYCL_ENABLE_HANDLER_FREE_SUBMISSION, since that is the main goal
(in a future MR). Made one CI build use this path, since the shims
we are using are fairly demanding on the programmer and compiler. 

Introduced supporting classes so that SYCL kernel-builder functions
can define storage local to a workgroup only one time, and then create
it where required.

Removed many uses of local_ptr<T> as these add little value beyond the
Hungarian-style sm_.* naming and are not easy to make from a 
work_group_static<T> object.

Binary reports the use of the extension.